The Summer addition
Now, we can have 0 courses in a semester. It is always assumed that you can have 0 courses in summer and winter. Winter is addable with 1 line of code if necessary. Summer should have a semester attribute "Summer" instead of the usual "Spring" and "Fall". This is the most verbose commit I've made. Nice.
This commit may not be stable and must be tested. It should work fine with the gui as it is.